refactor to include parent class and super class, all tests passing
